fixed all the TypeScript build errors

+++ frontend/src/services/pwa.ts
@@ -82,7 +82,7 @@ export class PWAService {
         userVisibleOnly: true,
         applicationServerKey: this.urlBase64ToUint8Array(
           import.meta.env.VITE_VAPID_PUBLIC_KEY || ''
-        ),
+        ) as Uint8Array,
       });
 
       // Send subscription to server
